refactor: 새로운 이미지 url을 입력하는 로직을 훅으로 변경
